requirements.txt
asn1crypto==1.2.0
astroid==2.3.3
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
conda==4.7.12
conda-package-handling==1.6.0
cryptography==2.8
Flask==1.1.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
menuinst==1.4.16
numpy==1.18.1
Pillow==7.0.0
pycosat==0.6.3
pycparser==2.19
pylint==2.4.4
pyOpenSSL==19.0.0
PySocks==1.7.1
python-dotenv==0.10.5
pywin32==223
requests==2.22.0
ruamel-yaml==0.15.46
six==1.12.0
torch==1.4.0+cpu
torchvision==0.5.0+cpu
tqdm==4.36.1
typed-ast==1.4.1
urllib3==1.24.2
Werkzeug==0.16.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2